requirements.txt
accelerate>=0.30
fastapi>=0.110
joblib>=1.3
kaggle>=1.6
lightgbm>=4.3
matplotlib>=3.8
numpy>=1.26
pandas>=2.2
plotly>=5.22
pydantic>=2.7
pytest>=8.2
python-multipart>=0.0.9
PyYAML>=6.0
scikit-learn>=1.4
scipy>=1.12
seaborn>=0.13
torch>=2.2
tqdm>=4.66
transformers>=4.41
uvicorn[standard]>=0.29
